+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 5.0.0
-- Added `cosmicCursor.smartDuplicate` command so it copies.
+- Added `cosmicCursor.smartDuplicate` command.
 
 ### 4.1.0
 - Amended `cosmicCursor.deleteRight` command so it does not spare a white-space before the next non-white-space character.